move rubrix client to proper file
